[benc/widget-props-13] Rename image.class.tsx to image.tsx
